Moved navbar components to own directory
